added extra step registration
