+++ vlinder/src/app/register/[accessToken]/@step_four/page.tsx
@@ -1,36 +1,24 @@
 "use client";
-import React from "react";
-import { useState, useEffect } from "react";
-
-//Enums
-import LookingFor, { LookingForDisplayNames } from "@/src/enums/LookingFor";
-
-//Components
-import { Input } from "@nextui-org/react";
-import { Button } from "@nextui-org/react";
-
-///backend
+import React, { useState, useEffect } from "react";
 import { useRouter, usePathname } from "next/navigation";
 import { createClient } from "@/utils/supabase/client";
-import { b, tr } from "framer-motion/client";
+import { Input, Button, Checkbox, Textarea } from "@nextui-org/react";
+
 const supabase = createClient();
 
-export default function Page() {
-  const [lookingFor, setLookingFor] = React.useState<LookingFor | undefined>(undefined);
+export default function PersonalInfoPage() {
+  const [smoker, setSmoker] = useState<boolean>();
+  const [disabilities, setDisabilities] = useState<string[]>([""]);
+  const [displayDisability, setDisplayDisability] = useState<boolean>();
   const [error, setError] = useState("");
   const [message, setMessage] = useState("");
   const router = useRouter();
   const pathName = usePathname();
-
-  const lookingForChange = (event: React.MouseEvent<HTMLButtonElement>) => {
-    const button = event.target as HTMLButtonElement;
-    const dataValue = button.getAttribute("data-gender") as LookingFor;
-    setLookingFor(dataValue);
-  };
+  const [loading, setLoading] = useState(true);
 
   const handleStepFourRegistration = async () => {
     const accessToken = pathName.split("/").pop();
-  
+
     const { data: tokenData, error: tokenError } = await supabase
       .from("accessToken")
       .select("*")
@@ -38,33 +26,32 @@ export default function Page() {
       .eq("is_used", false)
       .single();
 
-  
     if (tokenError || !tokenData) {
       router.push(`/register`);
       return;
     }
 
     const { data: userData, error: userError } = await supabase.auth.getUser();
-    
+
     if (userError || !userData) {
-      setMessage("PLease first enter your credentials");
+      setError("Please first enter your credentials");
       router.push(`/register/${accessToken}`);
       return;
     }
-    
+
     const { data: profileData, error: profileError } = await supabase
       .from("profiles")
       .select("*")
       .eq("id", userData.user.id)
       .single();
 
-
     if (!profileError && profileData) {
-      const lookingFor = profileData.sex_positive === true ? LookingFor.PartnerAndFriends : LookingFor.Friends;
-      setLookingFor(lookingFor);
+      setSmoker(profileData.smoker);
+      setDisabilities(profileData.disability || [""]);
+      setDisplayDisability(profileData.display_disability);
     }
 
-        
+    setLoading(false);
   };
 
   useEffect(() => {
@@ -75,12 +62,6 @@ export default function Page() {
     setError("");
     setMessage("");
 
-
-    if (lookingFor === null) {
-      setError("Please select your preference");
-      return;
-    }
-
     const { data: userData, error: userError } = await supabase.auth.getUser();
 
     if (!userData.user) {
@@ -88,45 +69,112 @@ export default function Page() {
       return;
     }
 
-    const sex_positive = lookingFor == "Friends" ? "false" : "true";
+    if (smoker === null || displayDisability === null) {
+        setError("Please fill in all fields.");
+        return;
+      }
 
     const { data, error } = await supabase
       .from("profiles")
       .upsert({
         id: userData.user.id,
-        sex_positive: sex_positive,
+        smoker: smoker,
+        disability: disabilities.filter((d) => d.trim() !== ""),
+        display_disability: displayDisability,
       });
 
     if (error) {
       setError("Error saving data");
       console.log(error);
       return;
-    }
-    else {
+    } else {
       setMessage("Information saved successfully");
     }
   };
 
+  const handleDisabilityChange = (index: number, value: string) => {
+    const newDisabilities = [...disabilities];
+    newDisabilities[index] = value;
+    setDisabilities(newDisabilities);
+  };
+
+  const addDisabilityField = () => {
+    setDisabilities([...disabilities, ""]);
+  };
+
+  const removeDisabilityField = (index: number) => {
+    const newDisabilities = disabilities.filter((_, i) => i !== index);
+    setDisabilities(newDisabilities);
+  };
+
   return (
     <section className="w-full flex flex-col justify-start items-center p-4">
-      <h2>What Are You Looking For?</h2>
+      <h2>Now, some personal questions</h2>
       <div className="w-full max-w-md p-8 h-fit">
-      <Button
-        color={lookingFor === LookingFor.Friends ? "primary" : "default"}
-        className="w-full mb-4 text-black"
-        data-gender={LookingFor.Friends}
-        onClick={lookingForChange}
-      >
-        {LookingForDisplayNames[LookingFor.Friends]}
-      </Button>
-      <Button
-        color={lookingFor === LookingFor.PartnerAndFriends ? "secondary" : "default"}
-        className="w-full mb-4 text-black"
-        data-gender={LookingFor.PartnerAndFriends}
-        onClick={lookingForChange}
-      >
-        {LookingForDisplayNames[LookingFor.PartnerAndFriends]}
-      </Button>
+      <div className="w-full mb-4 text-black">
+          <p>Are you a smoker?</p>
+          <div className="flex gap-4">
+            <Checkbox
+              isSelected={smoker === true}
+              onChange={() => setSmoker(true)}
+              color="secondary"
+            >
+              Yes
+            </Checkbox>
+            <Checkbox
+              isSelected={smoker === false}
+              onChange={() => setSmoker(false)}
+              color="secondary"
+            >
+              No
+            </Checkbox>
+          </div>
+        </div>
+        <div className="w-full mb-4 text-black">
+          <p>Disabilities</p>
+          {loading ? (
+            <p>Loading...</p>
+          ) : (
+            disabilities.map((disability: string, index: number) => (
+              <div key={index} className="flex items-center gap-2 mb-2">
+                <Input
+                  type="text"
+                  placeholder="Enter your disability"
+                  className="w-full"
+                  value={disability}
+                  onChange={(e) => handleDisabilityChange(index, e.target.value)}
+                />
+                {index > 0 && (
+                  <Button color="warning" onClick={() => removeDisabilityField(index)}>
+                    -
+                  </Button>
+                )}
+              </div>
+            ))
+          )}
+          <Button color="secondary" onClick={addDisabilityField}>
+            +
+          </Button>
+        </div>
+        <div className="w-full mb-4 text-black">
+            <p>Would you like to display your disability?</p>
+            <div className="flex gap-4">
+                <Checkbox
+                isSelected={displayDisability === true}
+                onChange={() => setDisplayDisability(true)}
+                color="secondary"
+                >
+                Yes
+                </Checkbox>
+                <Checkbox
+                isSelected={displayDisability === false}
+                onChange={() => setDisplayDisability(false)}
+                color="secondary"
+                >
+                No
+                </Checkbox>
+            </div>
+          </div>
         {error && <p className="text-red-500">{error}</p>}
         {message && <p className="text-green-500">{message}</p>}
         <Button className="w-full mb-4 text-black" color="primary" onClick={handleSave}>
