+++ vlinder/src/app/register/[accessToken]/@step_tree/page.tsx
@@ -59,7 +59,7 @@ export default function Page() {
     const { data: userData, error: userError } = await supabase.auth.getUser();
     
     if (userError || !userData) {
-      setMessage("PLease first enter your credentials");
+      setError("Please first enter your credentials");
       router.push(`/register/${accessToken}`);
       return;
     }
